grsec: Apply grsecurity-2.9.1-3.2.44-201304292055.patch
commit 3fbdd6071dbb0cee76b29eb12b7b79606a248a94
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Apr 29 18:44:23 2013 -0400

    add module.h to silence compiler warning, thanks to
    Sergei Trofimovich

 fs/btrfs/inode.c |    1 +
 1 files changed, 1 insertions(+), 0 deletions(-)

commit e0966e03b7fd8f9b47b039e2576fc159259738d2
Merge: 1ecc5df fe7a9b2
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Apr 29 18:54:49 2013 -0400

    Merge branch 'pax-stable2' into grsec-stable2

commit fe7a9b2587887771bd585d060916d19ff7f8a2da
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Apr 29 18:54:02 2013 -0400

    Update to pax-linux-3.2.44-test104.patch:
    - changed PHYSICAL_ALIGN/START to fix http://forums.grsecurity.net/viewtopic.php?f=3&t=3414
    - CONSTIFY depends on KERNEXEC (for the kernel open/close feature)
    - fixed CONSTIFY and powerpc interference, reported by  John Hardin (https://bugs.gentoo.org/show_bug.cgi?id=456364)
    - removed size overflow coverage of resource_size(), reported at http://forums.grsecurity.net/viewtopic.php?f=3&t=3412
    - fixed bad pointer arithmetic in nfsd_cache_update, reported by Jason A. Donenfeld and http://forums.grsecurity.net/viewtopic.php?f=3&t=3438
      note that the false positive is not fixed yet
    - fixed a few unintended bitmask computations found by a not-yet-public gcc plugin
    - fixed the kernel stack leak bug in do_tgkill, found by the size overflow plugin (https://code.google.com/p/chromium/issues/detail?id=223444)

 arch/x86/kernel/entry_64.S       |   11 ++++++++++-
 arch/x86/kernel/i8259.c          |    2 +-
 arch/x86/kernel/pci-calgary_64.c |    2 +-
 arch/x86/kvm/vmx.c               |    4 ++--
 drivers/block/pktcdvd.c          |    2 +-
 fs/btrfs/extent-tree.c           |    2 +-
 fs/nfsd/nfscache.c               |    6 ++++--
 kernel/trace/trace.c             |    2 +-
 kernel/trace/trace.h             |    2 +-
 tools/gcc/structleak_plugin.c    |    4 ++++
 10 files changed, 26 insertions(+), 11 deletions(-)

Signature-tree: abecb01ceceae118f414261ef20e0de4abadde0f

+++ arch/x86/kvm/vmx.c
@@ -1099,12 +1099,12 @@ static void vmcs_write64(unsigned long field, u64 value)
 #endif
 }
 
-static void vmcs_clear_bits(unsigned long field, u32 mask)
+static void vmcs_clear_bits(unsigned long field, unsigned long mask)
 {
 	vmcs_writel(field, vmcs_readl(field) & ~mask);
 }
 
-static void vmcs_set_bits(unsigned long field, u32 mask)
+static void vmcs_set_bits(unsigned long field, unsigned long mask)
 {
 	vmcs_writel(field, vmcs_readl(field) | mask);
 }
